# ui/desktop/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "goose-server",
    "description": "An AI agent",
    "contact": {
      "name": "Block",
      "email": "ai-oss-tools@block.xyz"
    },
    "license": {
      "name": "Apache-2.0"
    },
    "version": "1.9.0"
  },
  "paths": {
...
    "/sampling/pending": {
      "get": {
        "tags": [
          "super::routes::sampling"
        ],
        "operationId": "get_pending_requests",
        "responses": {
          "200": {
            "description": "List of pending sampling requests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SamplingRequest"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/sampling/stream": {
      "get": {
        "tags": [
          "super::routes::sampling"
        ],
        "operationId": "stream_sampling_requests",
        "responses": {
          "200": {
            "description": "SSE stream of sampling requests"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/sampling/{request_id}/approve": {
      "post": {
        "tags": [
          "super::routes::sampling"
        ],
        "operationId": "approve_sampling_request",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Approval recorded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalResponse"
                }
              }
            }
          },
          "404": {
            "description": "Request not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApprovalRequest": {
        "type": "object",
        "required": [
          "approved"
        ],
        "properties": {
          "approved": {
            "type": "boolean"
          }
        }
      },
      "ApprovalResponse": {
        "type": "object",
        "required": [
          "success"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "Content": {
        "oneOf": [
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/RawTextContent"
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/RawImageContent"
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/RawEmbeddedResource"
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/RawAudioContent"
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/RawResource"
              }
            ]
          }
        ]
      },
      "ContextInclusion": {
        "oneOf": [
          {
            "type": "string"
          },
          {
            "type": "string"
          },
          {
            "type": "string"
          }
        ]
      },
...
      "CreateMessageRequestParam": {
        "type": "object",
        "required": [
          "messages",
          "maxTokens"
        ],
        "properties": {
          "includeContext": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContextInclusion"
              },
              {
                "type": "object"
              }
            ]
          },
          "maxTokens": {
            "type": "integer",
            "minimum": 0
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SamplingMessage"
            }
          },
          "metadata": {
            "type": "object"
          },
          "modelPreferences": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelPreferences"
              },
              {
                "type": "object"
              }
            ]
          },
          "stopSequences": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "systemPrompt": {
            "type": "string"
          },
          "temperature": {
            "type": "number"
          }
        }
      },
...
      "Icon": {
        "type": "object",
        "required": [
          "src"
        ],
        "properties": {
          "mimeType": {
            "type": "string"
          },
          "sizes": {
            "type": "string"
          },
          "src": {
            "type": "string"
          }
        }
      },
...
      "ModelHint": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
      "ModelPreferences": {
        "type": "object",
        "properties": {
          "costPriority": {
            "type": "number"
          },
          "hints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelHint"
            }
          },
          "intelligencePriority": {
            "type": "number"
          },
          "speedPriority": {
            "type": "number"
          }
        }
      },
...
      "RawAudioContent": {
        "type": "object",
        "required": [
          "data",
          "mimeType"
        ],
        "properties": {
          "data": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          }
        }
      },
...
      "RawEmbeddedResource": {
        "type": "object",
        "required": [
          "resource"
        ],
        "properties": {
          "_meta": {
            "type": "object",
            "additionalProperties": true
          },
          "resource": {
            "$ref": "#/components/schemas/ResourceContents"
          }
        }
      },
      "RawImageContent": {
        "type": "object",
        "required": [
          "data",
          "mimeType"
        ],
        "properties": {
          "_meta": {
            "type": "object",
            "additionalProperties": true
          },
          "data": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          }
        }
      },
      "RawResource": {
        "type": "object",
        "required": [
          "uri",
          "name"
        ],
        "properties": {
          "description": {
            "type": "string"
          },
          "icons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Icon"
            }
          },
          "mimeType": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "uri": {
            "type": "string"
          }
        }
      },
      "RawTextContent": {
        "type": "object",
        "required": [
          "text"
        ],
        "properties": {
          "_meta": {
            "type": "object",
            "additionalProperties": true
          },
          "text": {
            "type": "string"
          }
        }
      },
...
      "ResourceContents": {
        "anyOf": [
          {
            "type": "object",
            "required": [
              "uri",
              "text"
            ],
            "properties": {
              "_meta": {
                "type": "object",
                "additionalProperties": true
              },
              "mimeType": {
                "type": "string"
              },
              "text": {
                "type": "string"
              },
              "uri": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "uri",
              "blob"
            ],
            "properties": {
              "_meta": {
                "type": "object",
                "additionalProperties": true
              },
              "blob": {
                "type": "string"
              },
              "mimeType": {
                "type": "string"
              },
              "uri": {
                "type": "string"
              }
            }
          }
        ]
      },
...
      "Role": {
        "oneOf": [
          {
            "type": "string"
          },
          {
            "type": "string"
          }
        ]
      },
...
      "SamplingMessage": {
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "properties": {
          "content": {
            "$ref": "#/components/schemas/Content"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        }
      },
      "SamplingRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CreateMessageRequestParam"
          },
          {
            "type": "object",
            "required": [
              "id",
              "extension_name"
            ],
            "properties": {
              "extension_name": {
                "type": "string"
              },
              "id": {
                "type": "string"
              }
            }
          }
        ]
      },
...
    }
  }
}